wildfire-detection-system/data_ingestion/ingest_data.py
from kafka import KafkaProducer
import json
import requests

producer = KafkaProducer(bootstrap_servers='kafka:9092', value_serializer=lambda v: json.dumps(v).encode('utf-8'))

def fetch_weather_data():
    response = requests.get('https://api.openweathermap.org/data/2.5/weather?q=CityName&appid=YOUR_API_KEY')
    data = response.json()
    producer.send('weather-data', data)

def fetch_satellite_data():
    # Placeholder function for fetching satellite imagery data
    pass

def fetch_vegetation_data():
    # Placeholder function for fetching vegetation indices data
    pass

fetch_weather_data()
fetch_satellite_data()
fetch_vegetation_data()




