wildfire-detection-system/data_ingestion/requirements.txt
kafka-python
requests
pymongo
